// src/app/Models/SetLeaveBO.ts
export class SetLeaveBO {
  EmployeeID: string;
  Enjoy: number;
  LeaveRemains: number;
  LeaveType: string;
  LeaveName: string;
  LeaveNo: string;
  EMPNO: string;
  LV_CAT_ID: string;
  App_No: string;
  APP_DATE: string;
  APV_BY: string;
  CHK: string;
  approved_day: number;
  RESON: string;
  CONTACT: string;
  LEAVE_STATUS: string;
  PERSON_ID: string;
  LEAVE_NAME: string;
  APV_DAYS: number;
  ApprovedBy: string;
  FromDate: string;
  ToDate: string;
  BALANCE: number;
  COM_ID: string;
  REMARKS: string;
}